+++ PolyCalc.h
@@ -5,6 +5,6 @@
 #ifndef LAGRANGE_INTERPOLATION_POLYCALC_H
 #define LAGRANGE_INTERPOLATION_POLYCALC_H
 
-int *polyCalc(int order, const int *org);
+long long *polyCalc(int order, const int *org);
 
 #endif //LAGRANGE_INTERPOLATION_POLYCALC_H
